Add dev registry style tests
